+++ pkg/formula/workspace/workspace_test.go
@@ -18,6 +18,8 @@ package workspace
 
 import (
 	"encoding/json"
+	"errors"
+	"fmt"
 	"io/ioutil"
 	"os"
 	"path"
@@ -29,6 +31,7 @@ import (
 
 	"github.com/ZupIT/ritchie-cli/internal/mocks"
 	"github.com/ZupIT/ritchie-cli/pkg/formula"
+	"github.com/ZupIT/ritchie-cli/pkg/formula/tree"
 	"github.com/ZupIT/ritchie-cli/pkg/stream"
 )
 
@@ -40,7 +43,9 @@ func TestWorkspaceManagerAdd(t *testing.T) {
 	_ = os.Mkdir(tmpDir, os.ModePerm)
 	defer os.RemoveAll(tmpDir)
 
-	dirManager := stream.NewDirManager(stream.NewFileManager())
+	fileManager := stream.NewFileManager()
+	dirManager := stream.NewDirManager(fileManager)
+	treeGen := tree.NewGenerator(dirManager, fileManager)
 	workspaceFile := path.Join(tmpDir, formula.WorkspacesFile)
 	workspaceBrokenPath := path.Join(tmpDir, "broken")
 	workspaceNonExistingPath := filepath.Join(tmpDir, "non-existing-dir")
@@ -113,7 +118,7 @@ func TestWorkspaceManagerAdd(t *testing.T) {
 			localBuilder := &mocks.LocalBuilderMock{}
 			localBuilder.On("Init", mock.AnythingOfType("string"), tt.workspace.Name).Return("", nil)
 
-			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder)
+			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder, treeGen)
 			got := workspace.Add(tt.workspace)
 
 			if got != nil {
@@ -143,7 +148,9 @@ func TestManagerDelete(t *testing.T) {
 
 	workspaceFile := path.Join(tmpDir, formula.WorkspacesFile)
 	fileNonExistentPath := path.Join(tmpDir, "non-existent")
-	dirManager := stream.NewDirManager(stream.NewFileManager())
+	fileManager := stream.NewFileManager()
+	dirManager := stream.NewDirManager(fileManager)
+	treeGen := tree.NewGenerator(dirManager, fileManager)
 
 	tests := []struct {
 		name          string
@@ -186,7 +193,7 @@ func TestManagerDelete(t *testing.T) {
 
 			localBuilder := &mocks.LocalBuilderMock{}
 
-			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder)
+			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder, treeGen)
 			got := workspace.Delete(tt.workspace)
 
 			if got != nil {
@@ -211,7 +218,9 @@ func TestManagerList(t *testing.T) {
 	_ = os.Mkdir(tmpDir, os.ModePerm)
 	defer os.RemoveAll(tmpDir)
 
-	dirManager := stream.NewDirManager(stream.NewFileManager())
+	fileManager := stream.NewFileManager()
+	dirManager := stream.NewDirManager(fileManager)
+	treeGen := tree.NewGenerator(dirManager, fileManager)
 
 	workspaceFile := path.Join(tmpDir, formula.WorkspacesFile)
 	err := ioutil.WriteFile(workspaceFile, []byte(`{"zup": "/some/path"}`), os.ModePerm)
@@ -248,7 +257,7 @@ func TestManagerList(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			localBuilder := &mocks.LocalBuilderMock{}
 
-			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder)
+			workspace := New(tt.workspacePath, tt.workspacePath, dirManager, localBuilder, treeGen)
 			got, err := workspace.List()
 
 			if err != nil {
@@ -261,8 +270,118 @@ func TestManagerList(t *testing.T) {
 	}
 }
 
+func TestWorkspaceManagerUpdate(t *testing.T) {
+	cleanForm()
+	fullDir := createFullDir()
+	userHome := os.TempDir()
+	ritHome := path.Join(userHome, ".rit-workspace-update")
+
+	workspaceFile := path.Join(userHome, formula.WorkspacesFile)
+
+	tests := []struct {
+		name          string
+		workspacePath string
+		workspace     formula.Workspace
+		outErr        string
+		treeGenErr    error
+	}{
+		{
+			name:          "success update",
+			workspacePath: ritHome,
+			workspace: formula.Workspace{
+				Name: "zup",
+				Dir:  fullDir,
+			},
+		},
+		{
+			name:          "list workspace error",
+			workspacePath: ritHome,
+			outErr:        "unexpected end of JSON input",
+		},
+		{
+			name:          "invalid workspace",
+			workspacePath: workspaceFile,
+			workspace: formula.Workspace{
+				Name: "commons",
+				Dir:  "home/user/go/src/github.com/ZupIT/ritchie-formulas-commons",
+			},
+			outErr: ErrInvalidWorkspace.Error(),
+		},
+		{
+			name:          "tree generation error",
+			workspacePath: ritHome,
+			workspace: formula.Workspace{
+				Name: "zup",
+				Dir:  fullDir,
+			},
+			treeGenErr: errors.New("error to generate tree.json"),
+			outErr:     "error to generate tree.json",
+		},
+		{
+			name:          "write error",
+			workspacePath: ritHome,
+			workspace: formula.Workspace{
+				Name: "commons",
+				Dir:  fullDir,
+			},
+			outErr: fmt.Sprintf(
+				"open %s: no such file or directory",
+				filepath.Join(ritHome, formula.ReposDir, "local-commons", tree.FileName),
+			),
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_ = os.Mkdir(ritHome, os.ModePerm)
+			defer os.RemoveAll(ritHome)
+
+			switch tt.name {
+			case "success update":
+				_ = os.Mkdir(path.Join(ritHome, formula.ReposDir), os.ModePerm)
+				_ = os.Mkdir(path.Join(ritHome, formula.ReposDir, "local-"+tt.workspace.Name), os.ModePerm)
+				err := ioutil.WriteFile(path.Join(ritHome, formula.ReposDir, "local-"+tt.workspace.Name, tree.FileName), []byte(`{"zup": "some/dir/path"}`), os.ModePerm)
+				assert.NoError(t, err)
+
+			case "list workspace error":
+				err := ioutil.WriteFile(path.Join(ritHome, formula.WorkspacesFile), []byte(``), os.ModePerm)
+				assert.NoError(t, err)
+			}
+
+			fileManager := stream.NewFileManager()
+			dirManager := stream.NewDirManager(fileManager)
+
+			localBuilder := &mocks.LocalBuilderMock{}
+			localBuilder.On("Init", mock.Anything, tt.workspace.Name).Return("", nil)
+
+			treeGen := &mocks.TreeManager{}
+			treeGen.On("Generate", mock.Anything).Return(formula.Tree{}, tt.treeGenErr)
+
+			workspaceManager := New(tt.workspacePath, userHome, dirManager, localBuilder, treeGen)
+			workspaceManager.Add(tt.workspace)
+
+			got := workspaceManager.Update(tt.workspace)
+
+			if got != nil {
+				assert.EqualError(t, got, tt.outErr)
+			} else {
+				assert.Empty(t, tt.outErr)
+				file, err := ioutil.ReadFile(path.Join(tt.workspacePath, formula.WorkspacesFile))
+				assert.NoError(t, err)
+				workspaces := formula.Workspaces{}
+				err = json.Unmarshal(file, &workspaces)
+				assert.NoError(t, err)
+				pathName := workspaces[tt.workspace.Name]
+				assert.Contains(t, tt.workspace.Dir, pathName)
+			}
+		})
+	}
+}
+
 func TestPreviousHash(t *testing.T) {
-	dirManager := stream.NewDirManager(stream.NewFileManager())
+	fileManager := stream.NewFileManager()
+	dirManager := stream.NewDirManager(fileManager)
+	treeGen := tree.NewGenerator(dirManager, fileManager)
 	tmpDir := os.TempDir()
 	ritHome := path.Join(tmpDir, ".rit")
 	defer os.RemoveAll(ritHome)
@@ -296,7 +415,7 @@ func TestPreviousHash(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			localBuilder := &mocks.LocalBuilderMock{}
 
-			workspace := New(tt.homePath, tt.homePath, dirManager, localBuilder)
+			workspace := New(tt.homePath, tt.homePath, dirManager, localBuilder, treeGen)
 			hash, err := workspace.PreviousHash(formulaPath)
 
 			if err != nil {
@@ -310,7 +429,9 @@ func TestPreviousHash(t *testing.T) {
 }
 
 func TestUpdateHash(t *testing.T) {
-	dirManager := stream.NewDirManager(stream.NewFileManager())
+	fileManager := stream.NewFileManager()
+	dirManager := stream.NewDirManager(fileManager)
+	treeGen := tree.NewGenerator(dirManager, fileManager)
 	ritHome := path.Join(os.TempDir(), "update-hash")
 	_ = os.Mkdir(ritHome, os.ModePerm)
 	defer os.RemoveAll(ritHome)
@@ -333,7 +454,7 @@ func TestUpdateHash(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			localBuilder := &mocks.LocalBuilderMock{}
 
-			workspace := New(tt.homePath, tt.homePath, dirManager, localBuilder)
+			workspace := New(tt.homePath, tt.homePath, dirManager, localBuilder, treeGen)
 			err := workspace.UpdateHash("my/formula", "hash")
 			assert.NoError(t, err)